first candidate for merge

+++ src/comp/Async.cpp
@@ -29,7 +29,7 @@ Async::Async(std::shared_ptr<base::Archetypes> archetypes, std::shared_ptr<Compo
 Async::~Async() { stop(); }
 
 void Async::run(size_t numberOfWorkerThreads) {
-    size_t workers = std::max(1ull, numberOfWorkerThreads);
+    size_t workers = std::max(static_cast<size_t>(1), numberOfWorkerThreads);
     spdlog::info("Async starting {} worker threads.", workers);
     for (size_t i = 0; i < workers; ++i) {
         std::string threadName = fmt::format("async_{}", i);
